demo/webapp/requirements.txt
Flask==0.10.1
tensorflow
tensorflow-serving-api
Pillow
gunicorn
numpy
nltk
tflearn
